BUGFIX: Variable capture for suspend didn't include arguments.
Now state variable capture includes arguments and omits the callback.
This is the correct way. We still have one step to go - which is to
make sure that the captures suspension is fully serializable and for the
parts that aren't, to figure out a way to recreate the non-serializable
parts. The right way here is like Scheme's `dynamic-wind`, but need
to think it through still.